requirements.txt
black==22.10.0
click==8.1.3
colorama==0.4.6
glfw==2.5.5
imageio==2.22.2
keyboard==0.13.5
mypy-extensions==0.4.3
Nuitka==1.1.6
numpy==1.23.4
ordered-set==4.1.0
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
PyOpenGL==3.1.6
tomli==2.0.1
